etl/requirements.txt
requests>=2.31.0
pandas>=2.1.0,<3.0.0
python-dotenv>=1.0.0
psycopg2-binary>=2.9.0
geopandas>=0.14.0
